+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Format impressions amount (thousand separators)
+### Fixed
+- License details url
 
 ## [1.18.0] - 2022-06-02
 ### Changed