+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Opportunity sort and filter tables in query parameters
+- Remembering selected filters when switching tabs
 
 ## [1.18.0] - 2022-06-02
 ### Changed