+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Opportunity sort and filter tables in query parameters
 - Remembering selected filters when switching tabs
 ### Fixed
-- Displaying an incorrect message about the lack of funds on the campaign page
+- Direct deal for metaverse
 
 ## [1.18.0] - 2022-06-02
 ### Changed