+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remembering selected filters when switching tabs
 ### Fixed
 - Direct deal for metaverse
+- Metamask connection
 
 ## [1.18.0] - 2022-06-02
 ### Changed