+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.18.2] - 2022-06-02
 ### Added
 - Opportunity sort and filter tables in query parameters
 - Remembering selected filters when switching tabs
@@ -433,7 +435,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Publisher features (Sites & AdUnits)
 
  
-[Unreleased]: https://github.com/adshares/adpanel/compare/v1.18.0...develop
+[Unreleased]: https://github.com/adshares/adpanel/compare/v1.18.2...develop
+[1.18.2]: https://github.com/adshares/adpanel/compare/v1.18.0...v1.18.2
 [1.18.0]: https://github.com/adshares/adpanel/compare/v1.17.1...v1.18.0
 [1.17.1]: https://github.com/adshares/adpanel/compare/v1.17.0...v1.17.1
 [1.17.0]: https://github.com/adshares/adpanel/compare/v1.16.1...v1.17.0